backend/requirements.txt
Flask==0.12.2
Flask-PyMongo
flask-login
python-dateutil
requests